+++ packages/ui/components/Viewer.tsx
@@ -1,11 +1,10 @@
 import React, { useRef, useState, useEffect, forwardRef, useImperativeHandle, useCallback } from 'react';
-import { createPortal } from 'react-dom';
 import Highlighter from 'web-highlighter';
 import hljs from 'highlight.js';
 import 'highlight.js/styles/github-dark.css';
 import { Block, Annotation, AnnotationType, EditorMode } from '../types';
 import { Frontmatter } from '../utils/parser';
-import { Toolbar } from './Toolbar';
+import { AnnotationToolbar } from './AnnotationToolbar';
 import { TaterSpriteSitting } from './TaterSpriteSitting';
 import { AttachmentsButton } from './AttachmentsButton';
 import { getIdentity } from '../utils/identity';
@@ -123,7 +122,7 @@ export const Viewer = forwardRef<ViewerHandle, ViewerProps>(({
   const modeRef = useRef<EditorMode>(mode);
   const onAddAnnotationRef = useRef(onAddAnnotation);
   const pendingSourceRef = useRef<any>(null);
-  const [toolbarState, setToolbarState] = useState<{ element: HTMLElement; source: any } | null>(null);
+  const [toolbarState, setToolbarState] = useState<{ element: HTMLElement; source: any; selectionText: string } | null>(null);
   const [hoveredCodeBlock, setHoveredCodeBlock] = useState<{ block: Block; element: HTMLElement } | null>(null);
   const [isCodeBlockToolbarExiting, setIsCodeBlockToolbarExiting] = useState(false);
   const [isCodeBlockToolbarLocked, setIsCodeBlockToolbarLocked] = useState(false);
@@ -434,8 +433,10 @@ export const Viewer = forwardRef<ViewerHandle, ViewerProps>(({
             window.getSelection()?.removeAllRanges();
           } else {
             // Show toolbar in selection mode
+            // Capture selection text now (preserves line breaks between blocks)
+            const selectionText = window.getSelection()?.toString() || source.text;
             pendingSourceRef.current = source;
-            setToolbarState({ element: doms[0] as HTMLElement, source });
+            setToolbarState({ element: doms[0] as HTMLElement, source, selectionText });
           }
         }
       }
@@ -695,16 +696,23 @@ export const Viewer = forwardRef<ViewerHandle, ViewerProps>(({
           )
         ))}
 
-        <Toolbar
-          highlightElement={toolbarState?.element ?? null}
-          onAnnotate={handleAnnotate}
-          onClose={handleToolbarClose}
-        />
+        {/* Text selection toolbar */}
+        {toolbarState && (
+          <AnnotationToolbar
+            element={toolbarState.element}
+            positionMode="center-above"
+            onAnnotate={handleAnnotate}
+            onClose={handleToolbarClose}
+            copyText={toolbarState.selectionText}
+            closeOnScrollOut
+          />
+        )}
 
         {/* Code block hover toolbar */}
         {hoveredCodeBlock && !toolbarState && (
-          <CodeBlockToolbar
+          <AnnotationToolbar
             element={hoveredCodeBlock.element}
+            positionMode="top-right"
             onAnnotate={handleCodeBlockAnnotate}
             onClose={handleCodeBlockToolbarClose}
             isExiting={isCodeBlockToolbarExiting}
@@ -1007,170 +1015,3 @@ const CodeBlock: React.FC<CodeBlockProps> = ({ block, onHover, onLeave, isHovere
   );
 };
 
-const CodeBlockToolbar: React.FC<{
-  element: HTMLElement;
-  onAnnotate: (type: AnnotationType, text?: string, imagePaths?: string[]) => void;
-  onClose: () => void;
-  isExiting: boolean;
-  onMouseEnter: () => void;
-  onMouseLeave: () => void;
-  onLockChange?: (locked: boolean) => void;
-}> = ({ element, onAnnotate, onClose, isExiting, onMouseEnter, onMouseLeave, onLockChange }) => {
-  const [step, setStep] = useState<'menu' | 'input'>('menu');
-  const [inputValue, setInputValue] = useState('');
-  const [imagePaths, setImagePaths] = useState<string[]>([]);
-  const [position, setPosition] = useState<{ top: number; right: number }>({ top: 0, right: 0 });
-  const inputRef = useRef<HTMLTextAreaElement>(null);
-
-  useEffect(() => {
-    if (step === 'input') inputRef.current?.focus();
-  }, [step]);
-
-  // Notify parent when locked (in input mode)
-  useEffect(() => {
-    onLockChange?.(step === 'input');
-  }, [step, onLockChange]);
-
-  // Update position on scroll/resize
-  useEffect(() => {
-    const updatePosition = () => {
-      const rect = element.getBoundingClientRect();
-      setPosition({
-        top: rect.top - 40,
-        right: window.innerWidth - rect.right,
-      });
-    };
-
-    updatePosition();
-    window.addEventListener('scroll', updatePosition, true);
-    window.addEventListener('resize', updatePosition);
-
-    return () => {
-      window.removeEventListener('scroll', updatePosition, true);
-      window.removeEventListener('resize', updatePosition);
-    };
-  }, [element]);
-
-  const { top, right } = position;
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (inputValue.trim() || imagePaths.length > 0) {
-      onAnnotate(AnnotationType.COMMENT, inputValue || undefined, imagePaths.length > 0 ? imagePaths : undefined);
-    }
-  };
-
-  return createPortal(
-    <div
-      className="annotation-toolbar fixed z-[100] bg-popover border border-border rounded-lg shadow-2xl"
-      style={{
-        top,
-        right,
-        animation: isExiting ? 'code-toolbar-out 0.15s ease-in forwards' : 'code-toolbar-in 0.2s ease-out',
-      }}
-      onMouseDown={e => e.stopPropagation()}
-      onMouseEnter={onMouseEnter}
-      onMouseLeave={onMouseLeave}
-    >
-      <style>{`
-        @keyframes code-toolbar-in {
-          from {
-            opacity: 0;
-            transform: translateY(12px);
-          }
-          to {
-            opacity: 1;
-            transform: translateY(0);
-          }
-        }
-        @keyframes code-toolbar-out {
-          from {
-            opacity: 1;
-            transform: translateY(0);
-          }
-          to {
-            opacity: 0;
-            transform: translateY(8px);
-          }
-        }
-      `}</style>
-      {step === 'menu' ? (
-        <div className="flex items-center p-1 gap-0.5">
-          <button
-            onClick={() => onAnnotate(AnnotationType.DELETION)}
-            title="Delete"
-            className="p-1.5 rounded-md transition-colors text-destructive hover:bg-destructive/10"
-          >
-            <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
-            </svg>
-          </button>
-          <button
-            onClick={() => setStep('input')}
-            title="Comment"
-            className="p-1.5 rounded-md transition-colors text-accent hover:bg-accent/10"
-          >
-            <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M7 8h10M7 12h4m1 8l-4-4H5a2 2 0 01-2-2V6a2 2 0 012-2h14a2 2 0 012 2v8a2 2 0 01-2 2h-3l-4 4z" />
-            </svg>
-          </button>
-          <div className="w-px h-5 bg-border mx-0.5" />
-          <button
-            onClick={onClose}
-            title="Cancel"
-            className="p-1.5 rounded-md transition-colors text-muted-foreground hover:bg-muted"
-          >
-            <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-      ) : (
-        <form onSubmit={handleSubmit} className="flex items-start gap-1.5 p-1.5 pl-3">
-          <textarea
-            ref={inputRef}
-            rows={1}
-            className="bg-transparent text-sm min-w-44 max-w-80 max-h-32 placeholder:text-muted-foreground resize-none px-2 py-1.5 focus:outline-none focus:bg-muted/30"
-            style={{ fieldSizing: 'content' } as React.CSSProperties}
-            placeholder="Add a comment..."
-            value={inputValue}
-            onChange={e => setInputValue(e.target.value)}
-            onKeyDown={e => {
-              if (e.key === 'Escape') setStep('menu');
-              // Enter to submit, Shift+Enter for newline
-              if (e.key === 'Enter' && !e.shiftKey) {
-                e.preventDefault();
-                if (inputValue.trim() || imagePaths.length > 0) {
-                  onAnnotate(AnnotationType.COMMENT, inputValue || undefined, imagePaths.length > 0 ? imagePaths : undefined);
-                }
-              }
-            }}
-          />
-          <AttachmentsButton
-            paths={imagePaths}
-            onAdd={(path) => setImagePaths(prev => [...prev, path])}
-            onRemove={(path) => setImagePaths(prev => prev.filter(p => p !== path))}
-            variant="inline"
-          />
-          <button
-            type="submit"
-            disabled={!inputValue.trim() && imagePaths.length === 0}
-            className="px-[15px] py-1 text-xs font-medium rounded bg-primary text-primary-foreground hover:opacity-90 disabled:opacity-50 transition-opacity self-stretch"
-          >
-            Save
-          </button>
-          <button
-            type="button"
-            onClick={() => setStep('menu')}
-            className="p-1 rounded text-muted-foreground hover:text-foreground hover:bg-muted transition-colors"
-          >
-            <svg className="w-3.5 h-3.5" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </form>
-      )}
-    </div>,
-    document.body
-  );
-};
